# openapi/switchboard.openapi.json
{
  "components": {
    "schemas": {
      "ClaimCompleteClaimCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for completing one claim.",
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "evidence": {
            "default": "",
            "title": "Evidence"
          },
          "final_status": {
            "default": "",
            "title": "Final Status",
            "type": "string"
          },
          "mission_project": {
            "default": "",
            "title": "Mission Project",
            "type": "string"
          },
          "project": {
            "default": "maxwell",
            "title": "Project",
            "type": "string"
          },
          "schema": {
            "default": "switchboard.claim.complete_claim_command.v1",
            "title": "Schema",
            "type": "string"
          }
        },
        "required": [
          "claim_id"
        ],
        "title": "CompleteClaimCommand",
        "type": "object"
      },
      "ClaimNextCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for scheduler claim_next.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "board_id": {
            "default": "",
            "title": "Board Id",
            "type": "string"
          },
          "capabilities": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "deliverable_id": {
            "default": "",
            "title": "Deliverable Id",
            "type": "string"
          },
          "idem_key": {
            "default": "",
            "title": "Idem Key",
            "type": "string"
          },
          "lanes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Lanes",
            "type": "array"
          },
          "max_budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Max Budget Usd"
          },
          "max_risk": {
            "default": "",
            "title": "Max Risk",
            "type": "string"
          },
          "milestone_id": {
            "default": "",
            "title": "Milestone Id",
            "type": "string"
          },
          "mission_id": {
            "default": "",
            "title": "Mission Id",
            "type": "string"
          },
          "override_identity_risk": {
            "default": false,
            "title": "Override Identity Risk",
            "type": "boolean"
          },
          "project": {
            "default": "maxwell",
            "title": "Project",
            "type": "string"
          },
          "require_work_session": {
            "default": false,
            "title": "Require Work Session",
            "type": "boolean"
          },
          "schema": {
            "default": "switchboard.claim.claim_next_command.v1",
            "title": "Schema",
            "type": "string"
          },
          "session_policy_profile": {
            "default": "",
            "title": "Session Policy Profile",
            "type": "string"
          },
          "ttl_seconds": {
            "default": 1800,
            "title": "Ttl Seconds",
            "type": "integer"
          },
          "work_session": {
            "additionalProperties": true,
            "title": "Work Session",
            "type": "object"
          },
          "work_session_id": {
            "default": "",
            "title": "Work Session Id",
            "type": "string"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "ClaimNextCommand",
        "type": "object"
      },
      "ClaimTaskCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for claiming one exact ready task.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "idem_key": {
            "default": "",
            "title": "Idem Key",
            "type": "string"
          },
          "override_identity_risk": {
            "default": false,
            "title": "Override Identity Risk",
            "type": "boolean"
          },
          "project": {
            "default": "maxwell",
            "title": "Project",
            "type": "string"
          },
          "require_work_session": {
            "default": false,
            "title": "Require Work Session",
            "type": "boolean"
          },
          "schema": {
            "default": "switchboard.claim.claim_task_command.v1",
            "title": "Schema",
            "type": "string"
          },
          "session_policy_profile": {
            "default": "",
            "title": "Session Policy Profile",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "ttl_seconds": {
            "default": 1800,
            "title": "Ttl Seconds",
            "type": "integer"
          },
          "work_session": {
            "additionalProperties": true,
            "title": "Work Session",
            "type": "object"
          },
          "work_session_id": {
            "default": "",
            "title": "Work Session Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "agent_id"
        ],
        "title": "ClaimTaskCommand",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contract-first REST surface for Switchboard. Request bodies for the paths below are generated from ``src/switchboard/contracts`` Pydantic models shared with MCP application commands.",
    "title": "Switchboard REST",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/txp/v1/claim_next": {
      "post": {
        "operationId": "claimNext",
        "parameters": [
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimNextCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Claim the next ready task",
        "tags": [
          "contracts"
        ]
      }
    },
    "/txp/v1/claim_task": {
      "post": {
        "operationId": "claimTask",
        "parameters": [
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimTaskCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Claim one exact ready task",
        "tags": [
          "contracts"
        ]
      }
    },
    "/txp/v1/complete_claim": {
      "post": {
        "operationId": "completeClaim",
        "parameters": [
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCompleteClaimCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Complete a claim and move work to In Review",
        "tags": [
          "contracts"
        ]
      }
    }
  }
}